Create SMD data models
An SMD must contain certain elements, but instead of baking that into the data models, I'll let the API handle it on export.
Turns out you can do type hinting in Python 2 with a comment like `# type: (int, List[str]) -> None | float`.
It took a lot of time, but I think this is very clean code.