Create SMD file API for export
Two utility functions because that stuff was being reused, and some lines were getting a bit long.
Using deepcopy for the SMDFileModel because it's mutable, and I make use of that by adding a node if there are none, a frame if there are none, and to each frame a bone if there are none.
I think this code turned out nicely as well.
I'll probably do import tomorrow.