Implement SMDEncoder
Back to a class based approach, but cleaner than last time.
Each function can be used on its own, and the class doesn't store the SMD or the output, but instead stores options (in this case just add_default_values).
I'm happy with how this turned out, even if it took hours of tweaking.