Make encoder methods private
It's kinda neat that you can use all of these independently, but I don't think I'll be able to do that for the decoder, and I don't want their features to be different.
Besides, not like I actually expect anyone to use anything other than the main encode / decode methods.
Oh and uh the order of some lines got a bit messed up last commit, so I'm fixing that in this commit.