Implement SMDDecoder
There's a glaring TODO, the variable for `_decode_number()` is kinda weird, and `decode_smd()` is a large top-down affair, but otherwise I think this is pretty nice.
I'm going to try to think of a way to divide it into more methods, perhaps sending the iterator and calling it `commands`.
The csv library works well, though it only separates by spaces, not tabs.
I'm not going to worry about that right now though, because tabs can cause StudioMDL to crash, so nobody uses them anyway.